+++ example_code/header_freq_tables.h
@@ -1,15 +1,18 @@
 #ifndef HEADER_FREQ_TABLES
 #define HEADER_FREQ_TABLES
 
+#include <stdint.h>
+
 #include <vector>
 #include <utility>
+
 using std::vector;
 using std::pair;
 
 struct FreqTables {
  public:
-  static vector<pair<unsigned int, long> > request_freq_table;
-  static vector<pair<unsigned int, long> > response_freq_table;
+  static vector<pair<uint16_t, uint32_t> > request_freq_table;
+  static vector<pair<uint16_t, uint32_t> > response_freq_table;
 };
 
 #endif
